Add transition to review detail
